Return hipStreamLegacy in default_stream
When hip_rocprim::default_stream was written, there was no hipStreamLegacy
handle. Instead, we returned hipStreamDefault. hipStreamDefault takes
on the behaviour of hipStreamLegacy when HIP_API_PER_THREAD_DEFAULT_STREAM
is defined, so returning it was safe as long as we checked that macro.

Now that hipStreamLegacy exists, there's no need to return hipStreamDefault
anymore. Returning hipStreamLegacy provides closer parity with CUDA, which
returns cudaStreamLegacy.